requirements.txt
flask==2.1.0
flask-sqlalchemy==3.0.2
flask-login==0.6.2
werkzeug==2.1.1
requests==2.26.0
redis==4.5.1
rq==1.11.1
pillow==9.3.0
pyjwt==2.6.0
